chore: move explore manager to right domain
