+++ src/modules/blockchainApplication/explore/components/BlockchainApplications/index.js
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 import React from 'react';
-import BlockchainApplicationStatistics from '../../../explore/components/BlockchainApplicationStatistics';
+import BlockchainApplicationStatistics from '../BlockchainApplicationStatistics';
 
 const ManageBlockchainAppications = ({ apps, statistics }) => (
   <BlockchainApplicationStatistics apps={apps} statistics={statistics} />
