+++ src/routes/routesMap.js
@@ -41,7 +41,7 @@ import SwitchAccount from '@account/components/SwitchAccount';
 import BackupRecoveryPhraseFlow from '@account/components/BackupRecoveryPhraseFlow';
 import RemoveCurrentAccountFlow from '@account/components/RemoveCurrentAccountFlow';
 import RemoveSelectedAccountFlow from '@account/components/RemoveSelectedAccountFlow';
-import BlockchainApplicationStatistics from '@blockchainApplication/manage/manager/ManageBlockchainApplications/BlockchainApplicationStatistics';
+import BlockchainApplicationStatistics from '@blockchainApplication/explore/manager/BlockchainApplicationStatistics';
 import BlockchainApplicationDetails from '@blockchainApplication/explore/components/BlockchainApplicationDetails';
 
 export default {
